modify youtube and exchange
