+++ youtube/config .ex.js
@@ -1,0 +1,6 @@
+const config = {
+  apiKey: "",
+  apiUrl: "",
+};
+
+export default config;
