+++ youtube/index.html
@@ -38,6 +38,6 @@
       </footer>
     </div>
     <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-    <script src="./js/main.js"></script>
+    <script type="module" src="./js/all.js"></script>
   </body>
 </html>
